type change improves gas efficiency (#73)
* type change improves gas efficiency

* delete note to self

+++ contracts/VM.sol
@@ -34,7 +34,7 @@ abstract contract VM {
       internal returns (bytes[] memory)
     {
         bytes32 command;
-        uint256 flags;
+        uint8 flags;
         bytes32 indices;
 
         bool success;
